requirements.txt
pandas==2.0.3
PyPDF2==3.0.1
crewai==0.28.2
python-dotenv==1.0.0
langchain-openai==0.0.5
requests==2.31.0
